Reworked the search endpoint to take a SearchQuery parameter
