+++ lib/meilisearch.dart
@@ -7,8 +7,8 @@ export 'src/facet_stat.dart';
 export 'src/match_position.dart';
 export 'src/searchable.dart';
 export 'src/multi_search_result.dart';
-export 'src/multi_search_query.dart';
-export 'src/search_query.dart';
+export 'src/query_parameters/multi_search_query.dart';
+export 'src/query_parameters/index_search_query.dart';
 export 'src/filter_builder/_exports.dart';
 export 'src/query_parameters/_exports.dart';
 export 'src/matching_strategy_enum.dart';
